Hopefully FINAL FIX for MP pallet bug
It looks like only pallets were affected by the bug anyway, not bales or logs
- only add objects when the game is MP
- only add objects to table if they are a "vehicle" (e.g. pallet)
- only fix objects when they are no longer moving
- clear table in case anything is added that is later sold

Note: it seems like GIANTS may be calling "link()" for the pallets on the client, since the fix involves linking to the map root node (clearing the existing links) on the client.  Calling the same code only on the server doesn't change the behaviour.